backend/requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-core
langchain-community
langchain-google-genai
